# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get a list of all branches",
                "operationId": "getAllBranches",
                "responses": {
                    "200": {
                        "description": "List of all branches"
                    }
                }
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "operationId": "storeBranch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Branch"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation errors)"
                    }
                }
            }
        },
        "/api/branches/{id}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get a branch by ID",
                "operationId": "getBranchById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
        "/api/branches/{id}/employees": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Add a new employee to a branch",
                "operationId": "storeEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "example": "Engineer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee added successfully"
                    },
                    "404": {
                        "description": "Branch not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation errors)"
                    }
                }
            }
        }
    }
}